drivers: ufs: schedule ice setup work in high priority workqueue
There are cases when inital ice setup failed, and a work entry needs
to be queued to reconfigure it. Block layer, however, is not aware of
this work and schedules its own retry work in kblockd workqueue.

The kblockd workqueue has high priority, and it blocks the ice setup
work. The former wastes a lot of CPU cycles meaninglessly while the
essential reconfigure work does not have a chance to run, and causes
significant stutter in UI.

This commit moves the reconfigure work to a high priority workqueue
so it has the same priority as retry work. Stutters are gone and
storage benchmarks show up to 20% improvement.