camera_v2: use vmalloc when allocating memory for init_focus data
This data block can be very large (up to 40KB on maple) and kmalloc
is likely to fail with this size when system memory is running low -
we don't have contiguous 40KB space, but plenty of 4KB pages.
vmalloc is well suited here, so just use it.

Note: only 90% allocations are 40KB, there are still some <4KB
requests when using front camera. Maybe we would a kvmalloc backport.